@@ -5,7 +5,6 @@ import android.content.pm.PackageManager;
 import android.content.res.Resources;
 import android.text.TextUtils;
 
-import org.apache.http.cookie.Cookie;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -24,25 +23,30 @@ import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 import java.lang.ref.WeakReference;
 import java.lang.reflect.Field;
+import java.lang.reflect.InvocationTargetException;
+import java.net.InetAddress;
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.net.URLEncoder;
 import java.net.UnknownHostException;
 import java.util.ArrayList;
 import java.util.Comparator;
+import java.util.HashMap;
 import java.util.Iterator;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Random;
 import java.util.regex.Pattern;
 
-import de.robv.android.xposed.XposedBridge;
-
 import static android.os.Looper.getMainLooper;
+import static de.robv.android.xposed.XposedBridge.log;
 import static de.robv.android.xposed.XposedHelpers.callMethod;
 import static de.robv.android.xposed.XposedHelpers.callStaticMethod;
 import static de.robv.android.xposed.XposedHelpers.findClass;
 
 @SuppressWarnings({"unused", "WeakerAccess"})
 public class Utility {
+    private static Map<String, InetAddress[]> dnsCache = new HashMap<>();
     private static SimpleResolver cnDnsResolver;
     private static WeakReference<Resources> moduleResources = new WeakReference<>(null);
 
@@ -63,10 +67,11 @@ public class Utility {
         boolean first = true;
         Iterator<String> keys = json.keys();
         while (keys.hasNext()) {
-            if (first)
+            if (first) {
                 first = false;
-            else
+            } else {
                 result.append("&");
+            }
 
             String key = keys.next();
             Object val = json.get(key);
@@ -78,14 +83,19 @@ public class Utility {
     }
 
     static String serialData(Map<String, String> dataMap) throws UnsupportedEncodingException {
+        if (dataMap == null) {
+            return "";
+        }
+
         StringBuilder result = new StringBuilder();
         boolean first = true;
 
         for (Map.Entry<String, String> entry : dataMap.entrySet()) {
-            if (first)
+            if (first) {
                 first = false;
-            else
+            } else {
                 result.append("&");
+            }
 
             result.append(URLEncoder.encode(entry.getKey(), "UTF-8"));
             result.append("=");
@@ -96,25 +106,35 @@ public class Utility {
     }
 
     @SuppressWarnings("deprecation")
-    static String serialCookies(List<Cookie> cookieList) throws UnsupportedEncodingException, JSONException {
+    static String serialCookies(List cookieList, Map<String, String> cookieMethods, String filterDomain) throws UnsupportedEncodingException {
         StringBuilder result = new StringBuilder();
         boolean first = true;
-        for (Cookie cookie : cookieList) {
-            if (first)
-                first = false;
-            else
-                result.append("; ");
+        for (Object cookie : cookieList) {
 
-            result.append(URLEncoder.encode(cookie.getName(), "UTF-8"));
-            result.append("=");
-            result.append(URLEncoder.encode(cookie.getValue(), "UTF-8"));
+            String domain = (String) callMethod(cookie, cookieMethods.get("domain"));
+            if (filterDomain == null || filterDomain.equals(domain)) {
+                if (first) {
+                    first = false;
+                } else {
+                    result.append("; ");
+                }
+
+                String name = (String) callMethod(cookie, cookieMethods.get("name"));
+                String value = (String) callMethod(cookie, cookieMethods.get("value"));
+
+                result.append(URLEncoder.encode(name, "UTF-8"));
+                result.append("=");
+                result.append(URLEncoder.encode(value, "UTF-8"));
+            }
         }
         return result.toString();
     }
 
+
     static String getIpByHost(String domain) throws UnknownHostException, TextParseException {
-        if (cnDnsResolver == null)
+        if (cnDnsResolver == null) {
             cnDnsResolver = new SimpleResolver(Settings.getDnsServer());
+        }
 
         // caches mechanism built-in, just look it up
         Lookup lookup = new Lookup(domain, Type.A);
@@ -128,23 +148,44 @@ public class Utility {
         }
     }
 
-    static String optString(JSONObject json, String key) {
-        // http://code.google.com/p/android/issues/detail?id=13830
-        if (json.isNull(key))
-            return null;
-        else
-            return json.optString(key, null);
+    public static InetAddress[] getIpByHostViaHttpDns(String domain) throws IOException, InvocationTargetException, IllegalAccessException, JSONException {
+        if (dnsCache.containsKey(domain)) {
+            return dnsCache.get(domain);
+        } else {
+            String raw = Http.get(String.format("http://119.29.29.29/d?dn=%s&ip=119.29.29.29", domain), false)
+                    .getResponseText();
+            String[] ss = raw.replaceAll("[ \r\n]", "").split(";");
+
+
+            InetAddress[] ips = new InetAddress[ss.length];
+            for (int i = 0; i < ss.length; i++) {
+                ips[i] = InetAddress.getByAddress(domain, InetAddress.getByName(ss[i]).getAddress());
+            }
+            dnsCache.put(domain, ips);
+            return ips;
+        }
     }
 
-    static Map<String, String> queryToMap(String dataString) {
-        Map<String, String> dataMap = new LinkedHashMap<>();
-        if (!TextUtils.isEmpty(dataString)) {
-            for (String s : dataString.split("&")) {
+    static String optString(JSONObject json, String key) {
+        // http://code.google.com/p/android/issues/detail?id=13830
+        if (json.isNull(key)) {
+            return null;
+        } else {
+            return json.optString(key, null);
+        }
+    }
+
+    static Map<String, String> combineRequestData(String path, Map<String, String> dataMap) throws URISyntaxException {
+        Map<String, String> map = dataMap != null ? dataMap : new HashMap<>();
+
+        String query = new URI(path).getQuery();
+        if (!TextUtils.isEmpty(query)) {
+            for (String s : query.split("&")) {
                 String[] data = s.split("=");
-                dataMap.put(data[0], data[1]);
+                map.put(data[0], data[1]);
             }
         }
-        return dataMap;
+        return map;
     }
 
     static Context getSystemContext() {
@@ -152,7 +193,7 @@ public class Utility {
         return (Context) callMethod(activityThread, "getSystemContext");
     }
 
-    static Resources getModuleResources() throws PackageManager.NameNotFoundException, IllegalAccessException {
+    static Resources getModuleResources() throws PackageManager.NameNotFoundException {
         Resources resources = moduleResources.get();
         if (resources == null) {
             resources = getSystemContext().createPackageContext(BuildConfig.APPLICATION_ID, Context.CONTEXT_IGNORE_SECURITY).getResources();
@@ -175,10 +216,11 @@ public class Utility {
                 boolean isFirstLine = true;
                 while ((line = input.readLine()) != null) {
                     sb.append(line);
-                    if (isFirstLine)
+                    if (isFirstLine) {
                         isFirstLine = false;
-                    else
+                    } else {
                         sb.append(System.getProperty("line.separator"));
+                    }
                 }
             } finally {
                 try {
@@ -186,7 +228,7 @@ public class Utility {
                         input.close();
                     }
                 } catch (IOException e) {
-                    XposedBridge.log(e);
+                    log(e);
                 }
             }
             return sb.toString();
@@ -210,31 +252,41 @@ public class Utility {
         return fs != null && fs.length > 0 ? fs[0] : null;
     }
 
-    static File[] findFiles(File dir, final String start, final String end, final int limit) {
-        if (dir != null && dir.exists() && dir.isDirectory())
+    static File[] findFiles(File dir, final String start, final String end, final Integer limit) {
+        if (dir != null && dir.exists() && dir.isDirectory()) {
             return dir.listFiles(new FilenameFilter() {
                 int find = 0;
 
                 @Override
                 public boolean accept(File file, String s) {
-                    if (find < limit
+                    if ((limit == null || find < limit)
                             && (TextUtils.isEmpty(start) || s.startsWith(start))
                             && (TextUtils.isEmpty(end) || s.endsWith(end))) {
                         find++;
                         return true;
-                    } else
+                    } else {
                         return false;
+                    }
                 }
             });
-        else {
+        } else {
             return null;
         }
     }
 
-    static boolean deleteFile(File file) {
-        return file != null && file.exists() && file.isFile() && file.delete();
+    static void deleteFile(File file) {
+        try {
+            //noinspection ResultOfMethodCallIgnored
+            file.delete();
+        } catch (Throwable ignored) {
+        }
     }
 
+    static void deleteFiles(File[] files) {
+        for (File file : files) {
+            deleteFile(file);
+        }
+    }
 
     static boolean containsField(Class source, String exact, String start, String end) {
         return findFirstField(source, exact, start, end) != null;
@@ -250,14 +302,16 @@ public class Utility {
         List<Field> returnFs = new ArrayList<>();
 
         for (Field f : fs) {
-            if (returnFs.size() == limit)
+            if (returnFs.size() == limit) {
                 break;
+            }
 
             String s = f.getType().getName();
             if ((TextUtils.isEmpty(exact) || s.equals(exact))
                     && (TextUtils.isEmpty(start) || s.startsWith(start))
-                    && (TextUtils.isEmpty(end) || s.endsWith(end)))
+                    && (TextUtils.isEmpty(end) || s.endsWith(end))) {
                 returnFs.add(f);
+            }
 
         }
         return returnFs.toArray(new Field[returnFs.size()]);
@@ -306,7 +360,7 @@ public class Utility {
         }
     }
 
-    static boolean isCallFromMyself() {
+    public static boolean isCallFromMyself() {
         StackTraceElement[] elements = Thread.currentThread().getStackTrace();
         boolean findAppStack = false;
         boolean findModStack = false;
@@ -331,6 +385,11 @@ public class Utility {
         new android.os.Handler(getMainLooper()).postDelayed(runnable, delay);
     }
 
+    public static int randInt(int min, int max) {
+        Random rand = new Random();
+        return rand.nextInt((max - min) + 1) + min;
+    }
+
     static class AlphanumComparator implements Comparator<String> {
         // http://www.davekoelle.com/files/AlphanumComparator.java
 
@@ -346,16 +405,18 @@ public class Utility {
             if (isDigit(c)) {
                 while (marker < slength) {
                     c = s.charAt(marker);
-                    if (!isDigit(c))
+                    if (!isDigit(c)) {
                         break;
+                    }
                     chunk.append(c);
                     marker++;
                 }
             } else {
                 while (marker < slength) {
                     c = s.charAt(marker);
-                    if (isDigit(c))
+                    if (isDigit(c)) {
                         break;
+                    }
                     chunk.append(c);
                     marker++;
                 }
@@ -363,6 +424,7 @@ public class Utility {
             return chunk.toString();
         }
 
+        @Override
         public int compare(String s1, String s2) {
             int thisMarker = 0;
             int thatMarker = 0;
@@ -377,7 +439,7 @@ public class Utility {
                 thatMarker += thatChunk.length();
 
                 int result;
-                if (isDigit(thisChunk.charAt(0)) && isDigit(thatChunk.charAt(0))) {
+                if (isDigit(thisChunk.charAt(0)) == isDigit(thatChunk.charAt(0))) {
                     int thisChunkLength = thisChunk.length();
                     result = thisChunkLength - thatChunk.length();
                     if (result == 0) {
@@ -392,8 +454,9 @@ public class Utility {
                     result = thisChunk.compareTo(thatChunk);
                 }
 
-                if (result != 0)
+                if (result != 0) {
                     return result;
+                }
             }
 
             return s1Length - s2Length;
