@@ -3,7 +3,6 @@ package bin.xposed.Unblock163MusicClient.ui;
 import android.annotation.SuppressLint;
 import android.app.AlertDialog;
 import android.content.ComponentName;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.SharedPreferences.OnSharedPreferenceChangeListener;
@@ -39,8 +38,9 @@ public class SettingsActivity extends PreferenceActivity implements OnSharedPref
 
 
     private void checkState() {
-        if (getActivatedModuleVersion() == -1)
+        if (getActivatedModuleVersion() == -1) {
             showNotActive();
+        }
     }
 
 
@@ -48,11 +48,7 @@ public class SettingsActivity extends PreferenceActivity implements OnSharedPref
         new AlertDialog.Builder(this)
                 .setCancelable(false)
                 .setMessage(R.string.hint_reboot_not_active)
-                .setPositiveButton(R.string.active_now, new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int id) {
-                        openXposed();
-                    }
-                })
+                .setPositiveButton(R.string.active_now, (dialog, id) -> openXposed())
                 .setNegativeButton(R.string.cancel, null)
                 .show();
     }
@@ -63,39 +59,41 @@ public class SettingsActivity extends PreferenceActivity implements OnSharedPref
             if (getPackageManager().queryIntentActivities(intent, 0).isEmpty()) {
                 intent = getPackageManager().getLaunchIntentForPackage("de.robv.android.xposed.installer");
             }
-            intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-                    .putExtra("section", "modules")
-                    .putExtra("fragment", 1)
-                    .putExtra("module", BuildConfig.APPLICATION_ID);
-            startActivity(intent);
+            if (intent != null) {
+                intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+                        .putExtra("section", "modules")
+                        .putExtra("fragment", 1)
+                        .putExtra("module", BuildConfig.APPLICATION_ID);
+                startActivity(intent);
+            }
         } else {
             Toast.makeText(this, R.string.xposed_installer_not_installed, Toast.LENGTH_SHORT).show();
         }
     }
 
+    private boolean isVXP() {
+        return System.getProperty("vxp") != null;
+    }
+
     private void checkIcon() {
-        if (Utility.isAppInstalled(this, "de.robv.android.xposed.installer")) {
-            final ComponentName aliasName = new ComponentName(this, SettingsActivity.this.getClass().getName() + "-Alias");
+        if (!isVXP() && Utility.isAppInstalled(this, "de.robv.android.xposed.installer")) {
+            final ComponentName aliasName = new ComponentName(this, SettingsActivity.this.getClass().getName() + "Alias");
             final PackageManager packageManager = getPackageManager();
             if (packageManager.getComponentEnabledSetting(aliasName) != PackageManager.COMPONENT_ENABLED_STATE_DISABLED) {
                 new AlertDialog.Builder(this)
                         .setCancelable(false)
                         .setMessage(R.string.hint_hide_icon)
-                        .setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
-                            public void onClick(DialogInterface dialog, int id) {
-                                packageManager.setComponentEnabledSetting(
-                                        aliasName,
-                                        PackageManager.COMPONENT_ENABLED_STATE_DISABLED,
-                                        PackageManager.DONT_KILL_APP);
-                            }
-                        }).show();
+                        .setPositiveButton(R.string.ok, (dialog, id) -> packageManager.setComponentEnabledSetting(
+                                aliasName,
+                                PackageManager.COMPONENT_ENABLED_STATE_DISABLED,
+                                PackageManager.DONT_KILL_APP)).show();
             }
         }
     }
 
 
     @SuppressWarnings({"deprecation", "ResultOfMethodCallIgnored"})
-    @SuppressLint("SetWorldReadable")
+    @SuppressLint({"SetWorldReadable", "WorldReadableFiles"})
     private void setWorldReadable() {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
             File dataDir = new File(getApplicationInfo().dataDir);
@@ -124,6 +122,8 @@ public class SettingsActivity extends PreferenceActivity implements OnSharedPref
         super.onPause();
         PreferenceManager.getDefaultSharedPreferences(this)
                 .unregisterOnSharedPreferenceChangeListener(this);
+
+        setWorldReadable();
     }
 
     @Override
